refactor : package 정리 (winning, lottery)
