+++ src/main/java/step2/winning/WinningResults.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.winning;
 
 import java.util.Map;
