+++ src/test/java/step2/LotteryTest.java
@@ -2,8 +2,9 @@ package step2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2.lottery.Lottery;
+import step2.lottery.LotteryNumber;
 
-import java.util.ArrayList;
 import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
